requirements.txt
requests>=2.31
Pillow>=10.0
beautifulsoup4>=4.12
PyYAML>=6.0
anthropic>=0.40
